daemon: split API/gRPC/HTTP server startup out of daemon_run.go
Pure code-motion refactor (#5661 cohort). Lifts the API-surface
bring-up and the per-transport commit-wiring seams into a sibling file
in package daemon:

  - the #5054/#5961 transport commit-wiring section comment
  - grpcCommitFn / grpcCommitConfirmedFn
  - restCommitFn / restCommitConfirmedFn
  - shellCommitFn / shellCommitConfirmedFn
  - startGRPCServer
  - startHTTPServer
  - resolveAPIBinds

Declarations moved verbatim; no rename, no reorder, no logic change.
The free-floating #5054/#5961 section comment that documents the six
commit-fn seams is carried across with them so the group stays
documented. Only the per-file import set is narrowed.

This drops daemon_run.go under the ~2000-LOC threshold.

Validation: go build ./..., go test ./pkg/daemon/... pass; gofmt clean.
