+++ pkg/daemon/daemon_run_servers.go
@@ -1,0 +1,573 @@
+package daemon
+
+import (
+	"context"
+	"log/slog"
+	"net"
+	"sync"
+
+	"github.com/psaab/xpf/pkg/api"
+	"github.com/psaab/xpf/pkg/config"
+	"github.com/psaab/xpf/pkg/eventengine"
+	"github.com/psaab/xpf/pkg/feeds"
+	"github.com/psaab/xpf/pkg/fwdstatus"
+	"github.com/psaab/xpf/pkg/grpcapi"
+	"github.com/psaab/xpf/pkg/ipmon"
+	"github.com/psaab/xpf/pkg/lldp"
+	"github.com/psaab/xpf/pkg/logging"
+	"github.com/psaab/xpf/pkg/rpm"
+	"github.com/psaab/xpf/pkg/webmgmt"
+)
+
+// #5054/#5961: transport commit-wiring seams.
+//
+// The gRPC, HTTP/REST, and local-shell commit handlers must ALL route an
+// operator commit through commitAndApplyOperator / commitConfirmedAndApplyOperator
+// so the peer-sync decision is derived from RG0 ownership
+// (rg0ConfigSyncAuthority) and is transport-independent — the #5054 invariant.
+// Before #5054 only gRPC synced the peer; a REST/shell commit left the standby
+// on stale config.
+//
+// These closures were inlined at each transport's construction site, so no test
+// could exercise the wiring: reverting one transport back to
+// commitAndApply(ctx, comment, false) left every test green (#5961). They are
+// extracted here as three named per-transport seams — one method pair per
+// transport — so each transport's wiring is independently reachable from a test
+// and independently fail-on-revert covered (configsync_transport_5054_test.go).
+// They are intentionally identical: the whole point of #5054 is that every
+// transport resolves peer-sync the same way; keeping a distinct named seam per
+// transport is what lets a single-transport regression turn a single test case
+// RED instead of staying silent.
+
+func (d *Daemon) grpcCommitFn() func(context.Context, string) (*config.Config, error) {
+	return func(ctx context.Context, comment string) (*config.Config, error) {
+		return d.commitAndApplyOperator(ctx, comment)
+	}
+}
+
+func (d *Daemon) grpcCommitConfirmedFn() func(context.Context, int) (*config.Config, error) {
+	return func(ctx context.Context, minutes int) (*config.Config, error) {
+		return d.commitConfirmedAndApplyOperator(ctx, minutes)
+	}
+}
+
+func (d *Daemon) restCommitFn() func(context.Context, string) (*config.Config, error) {
+	return func(ctx context.Context, comment string) (*config.Config, error) {
+		return d.commitAndApplyOperator(ctx, comment)
+	}
+}
+
+func (d *Daemon) restCommitConfirmedFn() func(context.Context, int) (*config.Config, error) {
+	return func(ctx context.Context, minutes int) (*config.Config, error) {
+		return d.commitConfirmedAndApplyOperator(ctx, minutes)
+	}
+}
+
+func (d *Daemon) shellCommitFn() func(context.Context, string) (*config.Config, error) {
+	return func(ctx context.Context, comment string) (*config.Config, error) {
+		return d.commitAndApplyOperator(ctx, comment)
+	}
+}
+
+func (d *Daemon) shellCommitConfirmedFn() func(context.Context, int) (*config.Config, error) {
+	return func(ctx context.Context, minutes int) (*config.Config, error) {
+		return d.commitConfirmedAndApplyOperator(ctx, minutes)
+	}
+}
+
+// startGRPCServer constructs and launches the gRPC API server goroutine.
+// Extracted verbatim from Run()'s PHASE 5 (#4662 Increment 2); the leaf
+// startup block carries no ordering dependency (same code, same call point).
+func (d *Daemon) startGRPCServer(ctx context.Context, wg *sync.WaitGroup, eventBuf *logging.EventBuffer, fwdSampler *fwdstatus.Sampler) {
+	// d.dp asserted against the local grpcDataPlane probe
+	// (runtime_probes.go) — structurally identical to
+	// pkg/grpcapi's package-private grpcRuntime
+	// (pkg/grpcapi/runtime.go, #1516/#1554). Go duck-types the
+	// assignment to grpcapi.Config.DP at this site; signature
+	// drift surfaces as a compile error here.
+	var grpcDP grpcDataPlane
+	if d.dp != nil {
+		if probe, ok := d.dp.(grpcDataPlane); ok {
+			grpcDP = probe
+		}
+	}
+	grpcSrv := grpcapi.NewServer(d.opts.GRPCAddr, grpcapi.Config{
+		Store:      d.store,
+		DP:         grpcDP,
+		EventBuf:   eventBuf,
+		GC:         d.gc,
+		Routing:    d.routing,
+		FRR:        d.frr,
+		IPsec:      d.ipsec,
+		Cluster:    d.cluster,
+		DHCP:       d.dhcp,
+		DHCPServer: d.dhcpServer,
+		RPMResultsFn: func() []*rpm.ProbeResult {
+			if d.rpm != nil {
+				return d.rpm.Results()
+			}
+			return nil
+		},
+		IPMonStatusFn: func() []ipmon.PolicyStatus {
+			if d.ipmon != nil {
+				return d.ipmon.Status()
+			}
+			return nil
+		},
+		// #2079: active NAT pool-utilization alarms for
+		// `show security alarms`.
+		NATPoolAlarmsFn: d.natPoolAlarms,
+		FeedsFn: func() map[string]feeds.FeedInfo {
+			if d.feeds != nil {
+				return d.feeds.AllFeeds()
+			}
+			return nil
+		},
+		// #3042: live feed-prefix overlay so the gRPC MatchPolicies
+		// simulator resolves feed-backed address-names to their live
+		// CIDRs, matching what the AF_XDP helper enforces.
+		FeedOverlayFn: func() map[string][]string {
+			return d.feedSnapshotsForConfig(d.store.ActiveConfig())
+		},
+		LLDPNeighborsFn: func() []*lldp.Neighbor {
+			if d.lldpMgr != nil {
+				return d.lldpMgr.Neighbors()
+			}
+			return nil
+		},
+		// #1387 inc-2: DHCP dynamic-DNS status sources for the
+		// `show ... dhcp-server dynamic-dns` ShowText topics.
+		DDNSStatsFn:          d.DDNSStats,
+		SurfaceADDNSStatsFn:  d.SurfaceAStats,
+		SurfaceADDNSStatusFn: d.SurfaceAStatus,
+		SurfaceADDNSForceFn:  d.ForceDDNSUpdate,
+		DDNSOwnedRecordsFn:   d.OwnedDDNSRecords,
+		// #2464: per-collector NetFlow v9 / IPFIX write-health for
+		// `show services flow-monitoring statistics`.
+		FlowCollectorHealthFn: d.FlowCollectorHealth,
+		// gRPC commits sync to the cluster peer atomically inside
+		// the apply lock so the peer can never observe an apply
+		// that hasn't yet been propagated. #5054: the RG0-ownership
+		// peer-sync decision now lives in commitAndApplyOperator,
+		// shared with the REST and local-shell commit paths so every
+		// transport converges identically. Wiring lives in the
+		// grpcCommitFn / grpcCommitConfirmedFn seams so it is
+		// fail-on-revert covered (#5961).
+		CommitFn:          d.grpcCommitFn(),
+		CommitConfirmedFn: d.grpcCommitConfirmedFn(),
+		// #5281: a gRPC zeroize runs the wipe under the SAME apply gate as
+		// commit/sync and enters a terminal reset generation so no concurrent
+		// or later config writer re-creates the erased state before the daemon
+		// stops. The grpcapi handler passes its performZeroizeWipe primitive.
+		ZeroizeFn: d.factoryReset,
+		VRRPMgr:   d.vrrpMgr,
+		RAMgr:     d.ra,
+		Version:   d.opts.Version,
+		FabricPeerAddrFn: func() []string {
+			var addrs []string
+			if d.syncPeerAddr != "" {
+				addrs = append(addrs, d.syncPeerAddr)
+			} else {
+				d.fabricMu.RLock()
+				if d.fabricPeerIP != nil {
+					addrs = append(addrs, d.fabricPeerIP.String())
+				}
+				d.fabricMu.RUnlock()
+			}
+			if d.syncPeerAddr1 != "" {
+				addrs = append(addrs, d.syncPeerAddr1)
+			} else {
+				d.fabricMu.RLock()
+				if d.fabricPeerIP1 != nil {
+					addrs = append(addrs, d.fabricPeerIP1.String())
+				}
+				d.fabricMu.RUnlock()
+			}
+			return addrs
+		},
+		FabricVRFDevice: func() string {
+			if c := d.store.ActiveConfig(); c != nil && c.Chassis.Cluster != nil {
+				cc := c.Chassis.Cluster
+				if cc.ControlInterface != "" || cc.FabricInterface != "" {
+					return "vrf-mgmt"
+				}
+			}
+			return ""
+		}(),
+		FwdSampler: fwdSampler,
+	})
+	d.grpcSrv = grpcSrv
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		if err := grpcSrv.Run(ctx); err != nil {
+			slog.Error("gRPC server error", "err", err)
+		}
+	}()
+	slog.Info("gRPC API server started", "addr", d.opts.GRPCAddr)
+}
+
+// startHTTPServer constructs and launches the HTTP REST API server goroutine
+// (bind-interface resolution, TLS, API auth, and loopback clamping). Extracted
+// verbatim from Run()'s PHASE 5 (#4662 Increment 3); a leaf startup block with
+// no ordering dependency (same code, same call point, still guarded by the
+// d.opts.APIAddr check in Run).
+func (d *Daemon) startHTTPServer(ctx context.Context, wg *sync.WaitGroup, eventBuf *logging.EventBuffer) {
+	// d.dp asserted against the local apiDataPlane probe
+	// (runtime_probes.go) — structurally identical to pkg/api's
+	// package-private apiRuntimeDataPlane. Go duck-types the
+	// assignment to api.Config.DP at this site; signature drift
+	// surfaces as a compile error here.
+	var apiDP apiDataPlane
+	if d.dp != nil {
+		if probe, ok := d.dp.(apiDataPlane); ok {
+			apiDP = probe
+		}
+	}
+	apiCfg := api.Config{
+		Addr:     d.opts.APIAddr,
+		Store:    d.store,
+		DP:       apiDP,
+		EventBuf: eventBuf,
+		GC:       d.gc,
+		Routing:  d.routing,
+		FRR:      d.frr,
+		IPsec:    d.ipsec,
+		DHCP:     d.dhcp,
+		VRRPMgr:  d.vrrpMgr,
+		// #5054: HTTP/REST commits sync to the peer on the SAME
+		// RG0-ownership policy as gRPC/local-shell. Peer convergence
+		// is transport-independent (decided in commitAndApplyOperator),
+		// so a REST commit no longer leaves the standby on stale config
+		// until an unrelated transport-disconnect reverse-sync. Wiring
+		// lives in the restCommitFn / restCommitConfirmedFn seams so it
+		// is fail-on-revert covered (#5961).
+		CommitFn:          d.restCommitFn(),
+		CommitConfirmedFn: d.restCommitConfirmedFn(),
+		// #758: surface compile state so /health returns 503
+		// when the dataplane has never compiled successfully.
+		CompileHealthFn: func() api.CompileHealthSnapshot {
+			h := d.CompileHealthSnapshot()
+			return api.CompileHealthSnapshot{
+				EverSucceeded:    h.EverSucceeded,
+				FailureCount:     h.FailureCount,
+				LastError:        h.LastError,
+				LastErrorUnixSec: h.LastErrorUnixSec,
+			}
+		},
+		// #4184: surface the day-0 / bootstrap config-import outcome so a
+		// failed import is visible on /health, not just a boot-time WARN.
+		BootstrapImportFn: func() api.BootstrapImportSnapshot {
+			b := d.BootstrapImportSnapshot()
+			return api.BootstrapImportSnapshot{
+				Status:  b.Status,
+				Error:   b.Error,
+				UnixSec: b.UnixSec,
+				Failed:  b.Failed,
+			}
+		},
+		// #1780 Path A: expose the per-phase age of the Go periodic
+		// neighbor-maintenance loop so a wedged guarded goroutine
+		// (stuck netlink/probe syscall) is observable as a climbing
+		// gauge before it manifests as the cold-connect hang.
+		NeighborPhaseAgeFn: d.NeighborPeriodicPhaseAges,
+		// #1880: 1 while the last applied FRR reload fell back to
+		// the additive vtysh -f path (stale-config removal deferred
+		// to the in-manager retry loop).
+		FRRReloadDegradedFn: func() bool {
+			if d.frr == nil {
+				return false
+			}
+			return d.frr.ReloadDegraded()
+		},
+		// #4899: 1 while the last DHCP-lease-change IPsec rebind
+		// failed and swanctl local_addrs are still bound to a stale
+		// lease address (the retry loop has not reconverged), so the
+		// operator sees a tunnel that cannot re-establish instead of a
+		// silently-dropped reload error.
+		IPsecRebindPendingFn: d.IPsecRebindPending,
+		// #3780: surface scheduler republish-failure so
+		// xpf_scheduler_republish_failed reads 1 (and
+		// xpf_scheduler_republish_stale_seconds climbs) while a
+		// scheduler window transition's republish has not converged —
+		// otherwise stale enforcement past the window is invisible to
+		// monitoring.
+		SchedulerRepublishFailedFn:       d.SchedulerRepublishFailed,
+		SchedulerRepublishStaleSecondsFn: d.SchedulerRepublishStaleSeconds,
+		// #5669: surface the bounded-age fail-closed escalation so
+		// xpf_scheduler_republish_fail_closed reads 1 once a persistently
+		// failing republish has crossed the bounded age and the scheduler has
+		// forced scheduled policies inactive (deny) — a crisp alarm distinct
+		// from the climbing stale-seconds age.
+		SchedulerRepublishFailClosedFn: d.SchedulerRepublishFailClosed,
+		// #1827: ip-monitoring policy state for the xpf_ipmon_*
+		// metric family.
+		IPMonStatusFn: func() []ipmon.PolicyStatus {
+			if d.ipmon != nil {
+				return d.ipmon.Status()
+			}
+			return nil
+		},
+		// #4423 L: ip-monitoring actuation-failure counter
+		// (xpf_ipmon_actuation_failures_total) — surfaces the silent
+		// #3757 self-heal retry loop so a degraded failover is visible.
+		IPMonActuationFailuresFn: func() uint64 {
+			if d.ipmon != nil {
+				return d.ipmon.ActuationFailures()
+			}
+			return 0
+		},
+		// #2157: event-options remediation action counters for the
+		// xpf_event_actions_* / xpf_event_action_queue_depth family.
+		EventActionStatsFn: func() eventengine.Stats {
+			if d.eventEngine != nil {
+				return d.eventEngine.Stats()
+			}
+			return eventengine.Stats{}
+		},
+		// #1895: currently-failed RPM probe-pin installs (tests
+		// holding state on ErrProbeSetup instead of probing the
+		// default path).
+		RPMPinFailedFn: func() float64 {
+			if d.rpm != nil {
+				return float64(d.rpm.PinInstallFailureCount())
+			}
+			return 0
+		},
+		// #1799: surface configstore persist-degraded state so
+		// /health returns 503 (and xpf_daemon_config_persist_degraded
+		// reads 1) while the running active config is not durable on
+		// disk (failed HA sync / auto-rollback persist, retry pending).
+		ConfigPersistDegradedFn: d.store.ConfigPersistDegraded,
+		// #3441: surface configstore rollback-history-degraded state so
+		// /health reports it (non-fatal) and
+		// xpf_config_rollback_persist_degraded reads 1 while the most
+		// recent commit failed to durably persist its text rollback
+		// files. The active config is durable; this flags a degraded
+		// recovery aid, so it does not 503.
+		RollbackHistoryDegradedFn: d.store.RollbackHistoryDegraded,
+		// #2050: surface dynamic-address feed staleness so the
+		// xpf_feed_seconds_since_last_success / xpf_feed_stale gauges
+		// read live status. A frozen enforced address set (retain-forever
+		// default) is otherwise invisible to monitoring.
+		FeedsFn: func() map[string]feeds.FeedInfo {
+			if d.feeds != nil {
+				return d.feeds.AllFeeds()
+			}
+			return nil
+		},
+		// #3042: live feed-prefix overlay so the REST match-policies
+		// simulator resolves feed-backed address-names to their live
+		// CIDRs, matching what the AF_XDP helper enforces.
+		FeedOverlayFn: func() map[string][]string {
+			return d.feedSnapshotsForConfig(d.store.ActiveConfig())
+		},
+		// #3104: live per-scheduler active-state so the REST
+		// match-policies simulator skips a scheduler-inactive policy
+		// exactly like the runtime, instead of returning a verdict the
+		// dataplane disagrees with. Sourced from the same daemon-local
+		// accessor the CLI/gRPC show surfaces use
+		// (Manager.PolicySchedulerActiveState via the dataplane adapter).
+		// ok=false when the dataplane is absent (NoDataplane); the simulator
+		// then fails closed and treats scheduled policies as inactive (#3414),
+		// matching the dataplane's nil-state behavior rather than certifying
+		// an as-if-active verdict it is skipping.
+		PolicySchedulerActiveStateFn: func() (map[string]bool, bool) {
+			if d.dp == nil {
+				return nil, false
+			}
+			p, ok := d.dp.(interface {
+				PolicySchedulerActiveState() map[string]bool
+			})
+			if !ok {
+				return nil, false
+			}
+			return p.PolicySchedulerActiveState(), true
+		},
+		// #1387 inc-2: DHCP dynamic-DNS counters for the
+		// xpf_dhcp_ddns_* metric family. Returns nil when the
+		// manager is absent (NoDataplane), omitting the family.
+		DDNSStatsFn:     d.DDNSStats,
+		SurfaceAStatsFn: d.SurfaceAStats,
+		// #2464: per-collector NetFlow v9 / IPFIX write-health for the
+		// xpf_flow_export_collector_* family + /services/flow-exporters.
+		FlowCollectorHealthFn: d.FlowCollectorHealth,
+		// #3747: per-exporter pending-batch queue depth / high-water /
+		// dropped-at-capacity count for the xpf_flow_export_batch_* family.
+		FlowExportBatchStatsFn: d.FlowExportBatchStats,
+		// #3419 M6: report whether this node is the active cluster member
+		// for RG0 so the REST session view's ha_active field matches the
+		// gRPC contract (server_sessions.go IsLocalPrimary(0)). Standalone
+		// (no cluster) reports active.
+		HAActiveFn: func() bool {
+			if d.cluster == nil {
+				return true
+			}
+			return d.cluster.IsLocalPrimary(0)
+		},
+		// #3423 M5: stamp this node's cluster id on the REST session
+		// list/summary/clear responses so a dashboard knows WHICH node it
+		// observed. Standalone (no cluster) reports node 0.
+		NodeIDFn: func() int {
+			if d.cluster == nil {
+				return 0
+			}
+			return d.cluster.NodeID()
+		},
+		// #3423 H5/M5: hand the REST session handlers the live gRPC
+		// server, which is HA-aware — its ClearSessions fans out the
+		// clear to the cluster peer (clearPeerSessions, x-peer-forwarded
+		// recursion guard) and its GetSessions/GetSessionSummary stamp the
+		// peer's table when include_peer is set. Resolved lazily so the
+		// closure can reference the gRPC server constructed below this
+		// block; guarded so a nil server resolves to a nil interface (not
+		// a non-nil typed-nil), keeping REST local-only until the gRPC
+		// server is up.
+		ClusterSessionFn: func() api.ClusterSessionService {
+			if d.grpcSrv == nil {
+				return nil
+			}
+			return d.grpcSrv
+		},
+	}
+	// #5866: the management HTTP/HTTPS listener is owned by
+	// managementReconciler. apiCfg here carries only the runtime deps; the
+	// reconciler re-derives the bind/port/TLS/auth from each ACTIVE config
+	// (resolveAPIBinds) so it can start the listener now AND, on every day-2
+	// commit, reconcile the live listener + authentication snapshot without a
+	// restart — make-before-break rebind on a bind/port/TLS change, live auth
+	// swap on an unchanged bind. Before #5866 the server was constructed once
+	// here and never reconciled, so a committed bind/TLS/port/auth change (e.g.
+	// a revoked credential) sat inert until a daemon restart.
+	d.mgmt = newManagementReconciler(d, apiCfg)
+	if err := d.mgmt.start(ctx); err != nil {
+		// A boot bind failure is non-fatal (matches the pre-#5866 async
+		// srv.Run error log): the daemon keeps running and the next commit's
+		// reconcileWebManagement retries the bind.
+		slog.Error("HTTP API server initial start failed", "err", err)
+	}
+	// Drain the management serve goroutines on daemon shutdown: ctx cancel
+	// triggers the api.Server bounded 5s graceful drain, and wait() joins every
+	// live + retiring listener goroutine so none leak past Run.
+	wg.Add(1)
+	go func() {
+		defer wg.Done()
+		<-ctx.Done()
+		d.mgmt.wait()
+	}()
+	slog.Info("HTTP API server started", "addr", d.opts.APIAddr)
+}
+
+// resolveAPIBinds finalizes the HTTP/HTTPS listen addresses, TLS flag, and
+// api-auth on apiCfg from the active web-management config, then applies the
+// #4047/#5127 loopback fail-safe clamp. It is split out of startHTTPServer so
+// the bind/auth/clamp decision is unit-testable without opening a socket.
+//
+// cfg is the active config; it may be nil or lack a `system services
+// web-management` stanza. The KEY invariant (#5127): the loopback clamp runs
+// UNCONDITIONALLY — even when no web-management block exists — because
+// apiCfg.Addr defaults to the `--api-addr` flag, which an operator can point
+// off-loopback with no web-management stanza at all. Before #5127 the clamp
+// lived INSIDE the web-management block, so that flag path bound the mutating
+// REST/config API (set / commit / rollback / DHCP / system-action) to the
+// network UNAUTHENTICATED, defeating the #4047 fail-safe.
+func (d *Daemon) resolveAPIBinds(apiCfg *api.Config, cfg *config.Config) {
+	if cfg != nil && cfg.System.Services != nil &&
+		cfg.System.Services.WebManagement != nil {
+		wm := cfg.System.Services.WebManagement
+		// #5715: an explicitly configured `web-management http` binds the
+		// canonical Junos J-Web port TCP/80 (webmgmt SSOT — the SAME port the
+		// host-inbound `http` service token admits, so the listener and the
+		// admission agree), on the configured interface address or loopback.
+		// With NO `web-management http` stanza (wm.HTTP false — e.g. an
+		// api-auth-only block) apiCfg.Addr is left untouched at the -api-addr
+		// default (127.0.0.1:8080), a loopback-only diagnostic path that is never
+		// host-inbound-admitted; that path must NOT be silently moved to port 80.
+		if wm.HTTP {
+			httpBindIP := "127.0.0.1"
+			if wm.HTTPInterface != "" {
+				httpBindIP = resolveInterfaceAddr(cfg, wm.HTTPInterface, "127.0.0.1")
+			}
+			// net.JoinHostPort (not string-concat) so an IPv6 interface
+			// address is bracketed ("[2001:db8::1]:80") — a bare
+			// "2001:db8::1:80" is unparseable by net.SplitHostPort (the
+			// clamp below) and net.Listen, blackholing an IPv6-only mgmt bind.
+			apiCfg.Addr = net.JoinHostPort(httpBindIP, webmgmt.HTTPPortString())
+			slog.Info("HTTP web-management bound", "interface", wm.HTTPInterface, "addr", apiCfg.Addr)
+		}
+		// Enable HTTPS if configured — bind the canonical TCP/443 (webmgmt SSOT).
+		if wm.HTTPS {
+			httpsBindIP := "127.0.0.1"
+			if wm.HTTPSInterface != "" {
+				httpsBindIP = resolveInterfaceAddr(cfg, wm.HTTPSInterface, "127.0.0.1")
+			}
+			apiCfg.TLS = true
+			apiCfg.HTTPSAddr = net.JoinHostPort(httpsBindIP, webmgmt.HTTPSPortString())
+			slog.Info("HTTPS web-management bound", "interface", wm.HTTPSInterface, "addr", apiCfg.HTTPSAddr)
+		}
+		// API authentication
+		if wm.APIAuth != nil && (len(wm.APIAuth.Users) > 0 || len(wm.APIAuth.APIKeys) > 0) {
+			authCfg := &api.AuthConfig{
+				Users:   make(map[string]string),
+				APIKeys: make(map[string]bool),
+			}
+			// #5636: never wire an EMPTY Basic password or empty api-key into
+			// the runtime AuthConfig. A quoted-empty secret parses as a real
+			// credential row but is not a valid credential — wiring it would let
+			// a request presenting `username:` (no password) or an empty
+			// Bearer / X-API-Key token authenticate, an auth bypass on an
+			// off-loopback bind. The commit gate rejects such a config, but an
+			// already-persisted / peer-synced config is lenient-loaded (#1960),
+			// so drop the empty credential here too (defense in depth; the
+			// middleware also rejects an empty configured secret).
+			for _, u := range wm.APIAuth.Users {
+				if pw := u.Password.Reveal(); pw != "" {
+					authCfg.Users[u.Username] = pw
+				}
+			}
+			for _, k := range wm.APIAuth.APIKeys {
+				if key := k.Reveal(); key != "" {
+					authCfg.APIKeys[key] = true
+				}
+			}
+			// Only enable auth when at least one USABLE credential survived; an
+			// all-empty api-auth stanza leaves apiCfg.Auth nil so the #4047
+			// runtime clamp still pulls a non-loopback bind back to loopback.
+			if len(authCfg.Users) > 0 || len(authCfg.APIKeys) > 0 {
+				apiCfg.Auth = authCfg
+				slog.Info("HTTP API authentication enabled", "users", len(authCfg.Users), "api_keys", len(authCfg.APIKeys))
+			} else {
+				slog.Warn("HTTP API api-auth configured with only empty secrets; ignoring (no valid credential) — #5636")
+			}
+		}
+	}
+
+	// #4047 part B / #5127: runtime fail-safe clamp, applied on EVERY path
+	// (whether or not a web-management stanza exists). The commit gate
+	// (validateWebManagementAuthStrict, pkg/config) rejects a NEW web-management
+	// config that binds the unauthenticated REST/config API off-loopback without
+	// api-auth, but an ALREADY-PERSISTED such config is lenient-loaded (warn, no
+	// brick — #1960), AND the `--api-addr` flag has no commit gate at all
+	// (#5127). Either can bind non-loopback UNAUTHENTICATED, exposing the
+	// mutating config endpoints (set / commit / rollback / DHCP / system-action)
+	// to the network. clampBindToLoopback is a no-op when the bind is already
+	// loopback OR api-auth is configured, so the default 127.0.0.1:8080 path and
+	// an authenticated off-loopback web-management bind are unaffected; it only
+	// pulls a non-loopback + no-auth bind back to a same-family loopback and
+	// WARNs. The daemon still boots and the web API stays reachable on loopback,
+	// with the console/SSH the lifeline (device-map §9.6 posture) until the
+	// operator adds api-auth (which restores the non-loopback bind).
+	hasAuth := apiCfg.Auth != nil
+	if clamped, ok := clampBindToLoopback(apiCfg.Addr, hasAuth); ok {
+		slog.Warn("HTTP API bind is non-loopback without api-auth; clamping to loopback (add `set system services web-management api-auth` to bind off-loopback) — #4047/#5127",
+			"requested", apiCfg.Addr, "clamped", clamped)
+		apiCfg.Addr = clamped
+	}
+	if apiCfg.TLS {
+		if clamped, ok := clampBindToLoopback(apiCfg.HTTPSAddr, hasAuth); ok {
+			slog.Warn("HTTPS API bind is non-loopback without api-auth; clamping to loopback — #4047/#5127",
+				"requested", apiCfg.HTTPSAddr, "clamped", clamped)
+			apiCfg.HTTPSAddr = clamped
+		}
+	}
+}
